# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyLabData",
    "version": "2.0.0.dev0"
  },
  "paths": {
...
    "/api/v1/imports/query": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "List Imports",
        "operationId": "list_imports_api_v1_imports_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/{import_id}/query": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Get Import",
        "operationId": "get_import_api_v1_imports__import_id__query_post",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/molecules/query": {
      "post": {
        "tags": [
          "molecules"
        ],
        "summary": "List Molecules",
        "operationId": "list_molecules_api_v1_molecules_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoleculeListQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/molecules/{molecule_id}/query": {
      "post": {
        "tags": [
          "molecules"
        ],
        "summary": "Get Molecule",
        "operationId": "get_molecule_api_v1_molecules__molecule_id__query_post",
        "parameters": [
          {
            "name": "molecule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Molecule Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/molecules/{molecule_id}/structure-3d/query": {
      "post": {
        "tags": [
          "molecules"
        ],
        "summary": "Get Molecule 3D Structure",
        "description": "Generate a 3D structure in memory; no result is persisted to DryData.",
        "operationId": "get_molecule_3d_structure_api_v1_molecules__molecule_id__structure_3d_query_post",
        "parameters": [
          {
            "name": "molecule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Molecule Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/molecules/{molecule_id}/attributes/query": {
      "post": {
        "tags": [
          "molecules"
        ],
        "summary": "List Molecule Attributes",
        "operationId": "list_molecule_attributes_api_v1_molecules__molecule_id__attributes_query_post",
        "parameters": [
          {
            "name": "molecule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Molecule Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/molecules/{molecule_id}/attributes/{attribute_id}/annotations/query": {
      "post": {
        "tags": [
          "molecules"
        ],
        "summary": "List Molecule Attribute Annotations",
        "operationId": "list_molecule_attribute_annotations_api_v1_molecules__molecule_id__attributes__attribute_id__annotations_query_post",
        "parameters": [
          {
            "name": "molecule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Molecule Id"
            }
          },
          {
            "name": "attribute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attribute Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attributes/{attribute_id}/entries/query": {
      "post": {
        "tags": [
          "attributes"
        ],
        "summary": "List Attribute Entries",
        "operationId": "list_attribute_entries_api_v1_attributes__attribute_id__entries_query_post",
        "parameters": [
          {
            "name": "attribute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attribute Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attributes/{attribute_id}/stats/query": {
      "post": {
        "tags": [
          "attributes"
        ],
        "summary": "Get Attribute Statistics",
        "operationId": "get_attribute_statistics_api_v1_attributes__attribute_id__stats_query_post",
        "parameters": [
          {
            "name": "attribute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attribute Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search Molecules",
        "operationId": "search_molecules_api_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/export": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Export Search",
        "description": "Submit an XLSX workbook export for one full search result set.",
        "operationId": "export_search_api_v1_search_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/molecules/{molecule_id}/properties/{entry_id}": {
      "put": {
        "tags": [
          "properties"
        ],
        "summary": "Update Property",
        "operationId": "update_property_api_v1_molecules__molecule_id__properties__entry_id__put",
        "parameters": [
          {
            "name": "molecule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Molecule Id"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "JobAccepted",
        "description": "Minimal response returned immediately after accepting a Job."
      },
...
      "MoleculeListQueryRequest": {
        "properties": {
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          },
          "query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          }
        },
        "type": "object",
        "title": "MoleculeListQueryRequest",
        "description": "Optional text filter and pagination for a molecule Query Job."
      },
...
      "PageQueryRequest": {
        "properties": {
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "PageQueryRequest",
        "description": "Pagination input for a Query Job that returns a collection."
      },
      "PropertyUpdateRequest": {
        "properties": {
          "value_number": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Number"
          },
          "value_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Text"
          },
          "value_boolean": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Boolean"
          },
          "source": {
            "type": "string",
            "minLength": 1,
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "PropertyUpdateRequest"
      },
      "SearchCondition": {
        "properties": {
          "attribute_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribute Id"
          },
          "attribute_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribute Key"
          },
          "entry_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "entry_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Key"
          },
          "operator": {
            "type": "string",
            "enum": [
              "eq",
              "ne",
              "lt",
              "lte",
              "gt",
              "gte",
              "between",
              "contains",
              "starts_with",
              "ends_with"
            ],
            "title": "Operator"
          },
          "value": {
            "title": "Value"
          },
          "second_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Second Value"
          }
        },
        "type": "object",
        "required": [
          "operator",
          "value"
        ],
        "title": "SearchCondition"
      },
      "SearchRequest": {
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/SearchCondition"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Conditions"
          },
          "logic": {
            "type": "string",
            "enum": [
              "and",
              "or"
            ],
            "title": "Logic",
            "default": "and"
          },
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "conditions"
        ],
        "title": "SearchRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}